requirements.txt
python-dotenv==1.0.0
instagrapi==2.0.0
Pillow==10.0.0
requests==2.31.0